+++ .idea/deploymentTargetDropDown.xml
@@ -7,11 +7,11 @@
         <deviceKey>
           <Key>
             <type value="VIRTUAL_DEVICE_PATH" />
-            <value value="C:\Users\PC_User\.android\avd\Pixel_6_API_30.avd" />
+            <value value="C:\Users\PC_User\.android\avd\Pixel_6_API_33.avd" />
           </Key>
         </deviceKey>
       </Target>
     </targetSelectedWithDropDown>
-    <timeTargetWasSelectedWithDropDown value="2023-05-11T15:01:12.419491800Z" />
+    <timeTargetWasSelectedWithDropDown value="2023-05-12T16:10:14.765027500Z" />
   </component>
 </project>
