+++ app/build.gradle
@@ -6,11 +6,11 @@ plugins {
 }
 
 android {
-    namespace 'com.example.okhttp'
+    namespace 'jp.search.okhttp'
     compileSdk 32
 
     defaultConfig {
-        applicationId "com.example.countrylist"
+        applicationId "jp.search.countrylist"
         minSdk 21
         targetSdk 32
         versionCode 1
