+++ app/src/main/AndroidManifest.xml
@@ -15,9 +15,9 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.Okhttp"
         tools:targetApi="31"
-        android:name="com.example.countrylist.SingletonContext">
+        android:name="jp.search.countrylist.SingletonContext">
         <activity
-            android:name="com.example.countrylist.view.Activity.MainActivity"
+            android:name="jp.search.countrylist.view.Activity.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
