--- app/src/main/java/com/example/countrylist/SingletonContext.kt
@@ -1,18 +1,0 @@
-package com.example.countrylist
-
-import android.app.Application
-import android.content.Context
-
-class SingletonContext : Application() {
-    init {
-        instance = this
-    }
-
-    companion object {
-        var instance: SingletonContext? = null
-
-        fun applicationContext(): Context? {
-            return instance!!.applicationContext
-        }
-    }
-}
