--- app/src/main/java/com/example/countrylist/model/DetailManager.kt
@@ -1,15 +1,0 @@
-package com.example.countrylist.model
-
-import android.widget.ImageView
-import com.example.countrylist.SingletonContext
-
-class DetailManager {
-    private val context = SingletonContext.applicationContext()!!
-    var item: String = ""
-
-    fun setPhoto(imageView: ImageView, engName: String?) {
-        item = engName + "img"
-        val pictureId = context.resources.getIdentifier(item,"drawable", context.packageName)
-        imageView.setImageResource(pictureId)
-    }
-}
