--- app/src/main/java/com/example/countrylist/model/Entity/Embassy.kt
@@ -1,6 +1,0 @@
-package com.example.countrylist.model.Entity
-
-data class Embassy(
-    var title: String,
-    var address: String
-)
