--- app/src/main/java/com/example/countrylist/model/Entity/Visa.kt
@@ -1,7 +1,0 @@
-package com.example.countrylist.model.Entity
-
-data class Visa(
-    var id: Int,
-    var name: String,
-    var content: String
-)
