--- app/src/main/java/com/example/countrylist/view/Activity/MainActivity.kt
@@ -1,16 +1,0 @@
-package com.example.countrylist.view.Activity
-
-import android.os.Bundle
-import androidx.appcompat.app.AppCompatActivity
-import com.example.okhttp.databinding.ActivityMainBinding
-
-class MainActivity : AppCompatActivity() {
-    private lateinit var binding: ActivityMainBinding
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        binding = ActivityMainBinding.inflate(layoutInflater)
-
-        setContentView(binding.root)
-    }
-}
