+++ app/src/main/java/jp/search/countrylist/model/Entity/Embassy.kt
@@ -1,0 +1,6 @@
+package jp.search.countrylist.model.Entity
+
+data class Embassy(
+    var title: String,
+    var address: String
+)
