+++ app/src/main/java/jp/search/countrylist/model/Entity/Visa.kt
@@ -1,0 +1,7 @@
+package jp.search.countrylist.model.Entity
+
+data class Visa(
+    var id: Int,
+    var name: String,
+    var content: String
+)
