+++ app/src/main/java/jp/search/countrylist/view/Activity/MainActivity.kt
@@ -1,0 +1,16 @@
+package jp.search.countrylist.view.Activity
+
+import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
+import jp.search.okhttp.databinding.ActivityMainBinding
+
+class MainActivity : AppCompatActivity() {
+    private lateinit var binding: ActivityMainBinding
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        binding = ActivityMainBinding.inflate(layoutInflater)
+
+        setContentView(binding.root)
+    }
+}
