+++ app/src/main/res/layout/activity_main.xml
@@ -7,6 +7,6 @@
         android:id="@+id/container"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:name="com.example.countrylist.view.Fragment.MainFragment" />
+        android:name="jp.search.countrylist.view.Fragment.MainFragment" />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
